+++ src/main/java/com/pismo/creditservice/vo/responses/TransactionResponse.java
@@ -22,7 +22,7 @@ public class TransactionResponse {
     @JsonProperty(value = "account_id")
     private Long accountId;
     @JsonProperty(value = "operation_type_id")
-    private Integer operationTypeId;
+    private Long operationTypeId;
     private BigDecimal amount;
 
 }
